backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sentence-transformers==3.0.1
transformers==4.41.2
torch==2.3.1
faiss-cpu==1.8.0
numpy==1.26.4
pdfplumber==0.11.1
pydantic==2.7.4
python-multipart==0.0.9
tqdm==4.66.4
